requirements.txt
hydra-core==1.3.2
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-openai==0.2.14
langchain-text-splitters==0.3.5
numpy==1.24.4
omegaconf==2.3.0
pandas==1.5.0
python-dotenv==1.0.1
streamlit==1.41.1
